experiments/mistralai-ocr/pyproject.toml
[project]
name = "mistra-ocr"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fire>=0.7.0",
    "mistralai>=1.7.0",
    "python-dotenv>=1.1.0",
]
